Refactor class inheritance using vtds-base provided base classes
